Fix event card and event detail ui bug
