+++ Event App/Event App/View/Home/HomeView.swift
@@ -13,6 +13,21 @@ struct HomeView : View {
     @State private var storedUserId: String = ""
     @State private var homeViewUsername: String = ""
     
+    @State var isLoading = false
+    let columns = [
+        GridItem(.adaptive(minimum: 165))
+    ]
+    
+    @State var events: [Events]?
+    
+    @State private var searchTerm = ""
+    @ObservedObject private var eventvm = EventViewModel()
+    
+    var filteredEvents: [Events] {
+        guard !searchTerm.isEmpty else {return events ?? []}
+        return events!.filter {$0.name.localizedCaseInsensitiveContains(searchTerm)}
+    }
+    
     init(){
         UITextField.appearance(whenContainedInInstancesOf: [UISearchBar.self]).backgroundColor = .white
         UITextField.appearance(whenContainedInInstancesOf: [UISearchBar.self]).tintColor = .black
@@ -39,8 +54,12 @@ struct HomeView : View {
                             endPoint: .bottom)
                     )
                     .edgesIgnoringSafeArea(.top)
+                    .zIndex(-2)
+
                 }
-                AllEventsViews()
+                
+                allEventsView
+                    .zIndex(0)
                                     
             }//end of ZStack
             .toolbar{
@@ -73,17 +92,26 @@ struct HomeView : View {
                 }
                 
             }
+            
+            .navigationBarBackButtonHidden(true)
+
         }
         .onAppear(perform: {
+            // Fetch event data
+            isLoading = true
+            fetchEventData()
+        })
+        .task {
+            // Get user data
             if let userId = UserDefaults.standard.string(forKey: "UserID"){
                 print("UserID in UserDefaults: \(userId)")
                 storedUserId = userId
                 userViewModel.fetchSingleUser(userId: userId) { result in
                     switch result {
                     case .success(let user):
-                        DispatchQueue.main.async {
-                            self.homeViewUsername = user.username
-                        }
+//                        DispatchQueue.main.async {
+                        self.homeViewUsername = user.username
+//                        }
                     case .failure(let error):
                         print(error.localizedDescription)
                     }
@@ -92,9 +120,55 @@ struct HomeView : View {
                 print("No UserID found in UserDefaults.")
                 storedUserId = ""
             }
-        })
-        .navigationBarBackButtonHidden(true)
+        }
+        
 
+    }
+}
+
+extension HomeView{
+    
+    
+    private var allEventsView: some View{
+        ScrollView(.vertical,showsIndicators: false) {
+            ZStack{
+                if isLoading {
+                    ProgressView("Loading...")
+                        .progressViewStyle(CircularProgressViewStyle())
+                        .tint(.white)
+                        .padding()
+                        .foregroundStyle(.white)
+                } else{
+                    LazyVGrid(columns: columns, spacing: 120) {
+                        
+                        ForEach(filteredEvents, id: \.id) { event in
+                            if event.isPending == false{
+                                EventCardView(event: .constant(event))
+                            }
+                        }
+                    }
+                }
+                
+            }
+            .padding(.bottom,70)
+            .padding(.top,60)
+            
+        }
+        .searchable(text: $searchTerm, placement: .navigationBarDrawer, prompt: "Search Events")
+
+    }
+    
+    func fetchEventData(){
+        eventvm.fetchEvents { result in
+            isLoading = false
+            switch result{
+            case .success(let events):
+                self.events = events
+            case .failure(let error):
+                print(error)
+            }
+        
+        }
     }
 }
